% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3horizon.R
\name{d3horizonOptions}
\alias{d3horizonOptions}
\title{Options for d3horizon}
\usage{
d3horizonOptions(height = 80, title = "", colors = c("#313695", "#4575b4",
  "#74add1", "#abd9e9", "#fee090", "#fdae61", "#f46d43", "#d73027"),
  mode = "mirror", ...)
}
\description{
Options for d3horizon
}
